Find config per-file, some optimizations.
Finding the config before `walk()` was incorrect; this is now slightly slower
but correct even if you specify a dir that is a parent of the one containing
pyproject.toml.  On Linux these duplicate loads are plenty fast, but we can
cache if this becomes an issue.

Adds `--show-time` option to CLI, which helped me find a bug that we were doing
versions in increasing order in `try_parse`; this shaves a 1.75 second run
(against `aql`) down to 1.3 seconds now.