Add configurable set of imports with side effects
When determining if imports are sortable, this adds some normalization,
and checks each part of the import name against the configured list
of known imports with side effects. If the full import, or any of the
combined stem parts, are in the configured list, then the import is
considered as unsortable, effectively creating an implicit barrier the
same way that `usort:skip` directives would.