Detect/enable the native LibCST parser by default
Adds a helper function to detect the presence of the `libcst.native`
parser module, and enable its use in LibCST appropriately. If the native
module is not available, or it does not contain the `parse_module`
function that we expect to use, then the native parser is not enabled.

Adds support for parsing 3.10 match statements, as well as future syntax
that requires use of the PEG parsing features only available in LibCST's
native parser.

This is currently only enabled in the µsort CLI entrypoint. Programs
using the µsort API directly will need to call or enable the native
parser separately if desired.

Fixes #216

[ghstack-poisoned]